EPFL-404 - Fix old entries cleaning (2018)

+++ data/wp/wp-content/plugins/epfl-404/inc/epfl-404-db.php
@@ -99,9 +99,23 @@ class EPFL404DB
     {
         global $wpdb;
 
+        /* We start by getting max ID to reuse it in delete request. We could use another way to do, for example
+         by using a subquery with "LIMIT" to select IDs to keep and delete oll the others but MariaDB doesn't
+         allows "LIMIT" in subqueries... */
+        $sql = "SELECT MAX(".self::EPFL404_DB_FIELD_ID.") AS 'max' FROM ".self::EPFL404_DB_TABLE;
+
+        if ( ($result = $wpdb->get_results( $sql )) === false )
+        {
+
+            /* We don't throw an exception this time, so website continue to works correctly, only 404 logging fails */
+            error_log("epfl-404: Error getting max id: ".$wpdb->print_error());
+            return;
+        }
+
+
         $sql = "DELETE FROM ".self::EPFL404_DB_TABLE.
                " WHERE ".self::EPFL404_DB_FIELD_DATE."< DATE_SUB(NOW(), INTERVAL ".$nb_days_to_keep." DAY)".
-               " OR ".self::EPFL404_DB_FIELD_ID."< (MAX(".self::EPFL404_DB_FIELD_ID.")-".$nb_entries_to_keep.")" ;
+               " OR ".self::EPFL404_DB_FIELD_ID."<= (".$result[0]->max."-".$nb_entries_to_keep.")" ;
 
         if ( $wpdb->query( $sql ) === false )
         {
